# Fetch line numbers for Python frames lazily