@godlygeek 2022-05-11
OK, I think this is now _actually_ correct... And the new test case covers the problem that the original version of this had, so that we know we won't regress on it.