@pablogsal 2022-08-17
Experiments executed for `python -m memray run  -o /dev/null -m test test_list test_list test_list test_list test_list test_list test_list test_list test_list test_list` for a total duration of 1.2595737380001992 seconds in baseline.

## With this PR

```
      Calls to PyFrame_GetLineNumber: 25550
     0.03%     0.03%  python  python3.10  [.] PyFrame_GetLineNumber                                                                                ▒
```

## Baseline


```
      Calls to PyFrame_GetLineNumber: 270090
     0.39%     0.39%  python  python3.10                               [.] PyFrame_GetLineNumber                                                   ◆
```

This means that in general the most we can save (that we are indeed saving it) is 0.39% of the total runtime (this is assuming no `PYTHONMALLOC=malloc` where most savings will be there since calls to `PyFrame_GetLineNumber` in `pushPythonFrame` are a bigger percentage of the runtime because our hooks are called less).

For the case where `PYTHONMALLOC=malloc` for a total duration of 3.0231499819997225 seconds in baseline:


```
      Calls to PyFrame_GetLineNumber: 21057170
+    5.26%     5.26%  python  python3.10                               [.] PyFrame_GetLineNumber
```

## Baseline

```
      Calls to PyFrame_GetLineNumber: 21125440
+    7.23%     7.23%  python  python3.10                               [.] PyFrame_GetLineNumber
```

This means that in the latter we are saving ~2% maximum. This is consistent with the timing results of `2.8712007330000233` (PR) against `2.935535135000009` (baseline).

## PR

```
  Test duration: 2.8712007330000233
  Time (mean ± σ):      3.058 s ±  0.051 s    [User: 3.008 s, System: 0.056 s]
  Range (min … max):    2.983 s …  3.167 s    10 runs
```

## Baseline

```
    Test duration: 2.935535135000009
  Time (mean ± σ):      3.120 s ±  0.046 s    [User: 3.070 s, System: 0.056 s]
  Range (min … max):    3.064 s …  3.210 s    10 runs
```

That is around 3.5% of the difference.